src/requirements.txt
PyGithub
cloudshell-shell-core>=2.0.0,<2.1.0
cloudshell-automation-api>=8.2.0,<8.3
